[FEAT] NFT 판매 등록 컨트랙트, 판매 종료 API 연동 (#S08P22D208-413, S08P22D208-408)

See merge request s08-blockchain-contract-sub2/S08P22D208!157

+++ Android/app/src/main/java/com/ssafy/kkaddak/data/remote/datasource/market/NftDetailItemResponse.kt
@@ -4,6 +4,7 @@ import com.google.gson.annotations.SerializedName
 import com.ssafy.kkaddak.data.remote.datasource.base.DataToDomainMapper
 import com.ssafy.kkaddak.domain.entity.market.HistoryItem
 import com.ssafy.kkaddak.domain.entity.market.NftDetailItem
+import java.math.BigInteger
 
 data class NftDetailItemResponse(
     @SerializedName("nftImageUrl")
@@ -17,7 +18,7 @@ data class NftDetailItemResponse(
     @SerializedName("isSelling")
     val isSelling: Boolean,
     @SerializedName("price")
-    val price: Double,
+    val price: BigInteger,
     @SerializedName("isLike")
     val isLike: Boolean,
     @SerializedName("sellerNickname")
