+++ Android/app/src/main/java/com/ssafy/kkaddak/domain/entity/market/HistoryItem.kt
@@ -1,6 +1,8 @@
 package com.ssafy.kkaddak.domain.entity.market
 
+import java.math.BigInteger
+
 data class HistoryItem(
-    val price: Double,
+    val price: BigInteger,
     val timestamp: String
 )
